Refactor cmd registration and submission, to make it more flexible.
Instead of having a strict definition of $namespace and $subcmd, allow
commands with "different" levels of specificity to be registered.

Allow both "foo" and "foo bar" to be registered, and invoke the "foo bar"
handler for all commands that start with the words "foo bar". Other commands,
such as "foo", "foo baz" or even "foo barf" are sent to the "foo" handles.
Finally, commands such as "bar" of "foos bar baz" are sent to the catch-all
handler registered for "" (the empty string).

As a consequence of making command handler registration more flexible, we can
now register complete commands, instead of just namespaces (i.e. instead of
registering "avr", and then having to continue parsing within AVR_Device's
handle_cmd() method, we can instead register "avr on", "avr off", etc. with
the same command handler, and it will never be invoked for commands it does
not know).

Registering full commands allow a future patch to AV_HTTPServer to dynamically
deduce the set of available commands, by traversing self.av_loop.cmd_handlers.