Make sure the first volume adjustment actually changes the volume
On the AVR, the first volume command does not change the volume, but rather
causes the current volume to be displayed for 2.5 seconds. Any subsequent
volume command within that time window will actually adjust the volume.

In order to make the volume commands always adjust the volume, we need to
keep track of whether the volume is currently being displayed or not.
This is done by the AVR_State.showing_volume attribute added here. It is
set to True/False (by AVR_State.update) based on whether the last status
update contained volume information or not.

The other part of the patch is in AVR_Device.handle_cmd: If the volume is not
currently shown when we're about schedule a volume command, then schedule
another copy of the same command to be sent immediately following the first.
The first will trigger the volume display, and the second will do the actual
adjustment. Also set AVR_State.showing_volume to True, when sending this
double command, so that if more volume commands are scheduled before the next
status update, we still only do the copying on the _first_ message.

Also, experiments show that the minimum inter-command timeout should be 0.25s.

Finally, AVR_State sometimes sends a volume command in order to just trigger
the volume display, so that it can learn the current volume. This command
should NOT adjust the volume, so introduce the "vol?" command, which is
identical to "vol-", except that it WON'T trigger the copying described above.