Revamp reading of AVR datagrams from the serial port.
The old code seems to depend on some coincidental synchronicity when reading
AVR status updates. However, on the Raspberry Pi, any read of 58 bytes from
the serial port does not automagically end up producing exactly one complete
datagram. Fix this by maintaining a read buffer, where we stuff the input from
the serial port that has not been parsed yet. Then skip bytes until the start
of a datagram is found, and keep readning until the read buffer contains the
entire datagram, and finally pass it on for proper parsing.